cpu::arm: Use `cfg_if!` to clarify target conditionals.
Clarify the conditions in which each implementation is used. Move each
OS-specific implementation into its own source file.

To review:
```
git difftool HEAD^1:src/cpu/arm.rs src/cpu/arm/darwin.rs
git difftool HEAD^1:src/cpu/arm.rs src/cpu/arm/linux.rs
git difftool HEAD^1:src/cpu/arm.rs src/cpu/arm/fuchsia.rs
git difftool HEAD^1:src/cpu/arm.rs src/cpu/arm/windows.rs
```
