+++ src/cpu/arm.rs
@@ -1,4 +1,4 @@
-// Copyright 2016-2021 Brian Smith.
+// Copyright 2016-2024 Brian Smith.
 //
 // Permission to use, copy, modify, and/or distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -40,199 +40,26 @@ mod abi_assumptions {
 // what target features are supported, so rely only on static feature
 // detection.
 
-#[cfg(all(
-    any(target_os = "android", target_os = "linux"),
-    not(target_env = "uclibc")
-))]
-fn detect_features() -> u32 {
-    cfg_if::cfg_if! {
-        if #[cfg(target_arch = "aarch64")] {
-            use libc::{getauxval,AT_HWCAP, HWCAP_AES, HWCAP_PMULL, HWCAP_SHA2, HWCAP_SHA512};
-        } else if #[cfg(target_arch = "arm")] {
-            // The `libc` crate doesn't provide this functionality on all
-            // 32-bit Linux targets, like Android or -musl. Use this polyfill
-            // for all 32-bit ARM targets so that testing on one of them will
-            // be more meaningful to the others.
-            use libc::c_ulong;
-            extern "C" {
-                pub fn getauxval(type_: c_ulong) -> c_ulong;
-            }
-            const AT_HWCAP: c_ulong = 16;
-            const AT_HWCAP2: c_ulong = 26;
-            const HWCAP_NEON: c_ulong = 1 << 12;
-            const HWCAP2_AES: c_ulong = 1 << 0;
-            const HWCAP2_PMULL: c_ulong = 1 << 1;
-            const HWCAP2_SHA2: c_ulong = 1 << 3;
-        }
-    }
-
-    let caps = unsafe { getauxval(AT_HWCAP) };
-
-    let mut features = 0;
-
-    // We do not need to check for the presence of NEON, as Armv8-A always has it
-    #[cfg(target_arch = "aarch64")]
-    const _ASSERT_NEON_DETECTED: () = assert!((ARMCAP_STATIC & NEON.mask) == NEON.mask);
-
-    // OpenSSL and BoringSSL don't enable any other features if NEON isn't
-    // available.
-    #[cfg(target_arch = "arm")]
-    if caps & HWCAP_NEON == HWCAP_NEON {
-        features |= NEON.mask;
+cfg_if::cfg_if! {
+    if #[cfg(all(target_arch = "aarch64",
+                 any(target_os = "ios", target_os = "macos", target_os = "tvos")))] {
+        mod darwin;
+        use darwin as detect;
+    } else if #[cfg(all(target_arch = "aarch64", target_os = "fuchsia"))] {
+        mod fuchsia;
+        use fuchsia as detect;
+    } else if #[cfg(any(target_os = "android", target_os = "linux"))] {
+        mod linux;
+        use linux as detect;
+    } else if #[cfg(all(target_arch = "aarch64", target_os = "windows"))] {
+        mod windows;
+        use windows as detect;
     } else {
-        return 0;
-    }
-
-    #[cfg(target_arch = "arm")]
-    let caps = unsafe { getauxval(AT_HWCAP2) };
-
-    #[cfg(target_arch = "arm")]
-    use {HWCAP2_AES as HWCAP_AES, HWCAP2_PMULL as HWCAP_PMULL, HWCAP2_SHA2 as HWCAP_SHA2};
-
-    if caps & HWCAP_AES == HWCAP_AES {
-        features |= AES.mask;
-    }
-    if caps & HWCAP_PMULL == HWCAP_PMULL {
-        features |= PMULL.mask;
-    }
-    if caps & HWCAP_SHA2 == HWCAP_SHA2 {
-        features |= SHA256.mask;
-    }
-
-    #[cfg(target_arch = "aarch64")]
-    if caps & HWCAP_SHA512 == HWCAP_SHA512 {
-        features |= SHA512.mask;
-    }
-
-    features
-}
-
-#[cfg(all(target_os = "fuchsia", target_arch = "aarch64"))]
-fn detect_features() -> u32 {
-    type zx_status_t = i32;
-
-    #[link(name = "zircon")]
-    extern "C" {
-        fn zx_system_get_features(kind: u32, features: *mut u32) -> zx_status_t;
-    }
-
-    const ZX_OK: i32 = 0;
-    const ZX_FEATURE_KIND_CPU: u32 = 0;
-    const ZX_ARM64_FEATURE_ISA_ASIMD: u32 = 1 << 2;
-    const ZX_ARM64_FEATURE_ISA_AES: u32 = 1 << 3;
-    const ZX_ARM64_FEATURE_ISA_PMULL: u32 = 1 << 4;
-    const ZX_ARM64_FEATURE_ISA_SHA2: u32 = 1 << 6;
-
-    let mut caps = 0;
-    let rc = unsafe { zx_system_get_features(ZX_FEATURE_KIND_CPU, &mut caps) };
-
-    let mut features = 0;
-
-    // We do not need to check for the presence of NEON, as Armv8-A always has it
-    const _ASSERT_NEON_DETECTED: () = assert!((ARMCAP_STATIC & NEON.mask) == NEON.mask);
-
-    if rc == ZX_OK {
-        if caps & ZX_ARM64_FEATURE_ISA_AES == ZX_ARM64_FEATURE_ISA_AES {
-            features |= AES.mask;
-        }
-        if caps & ZX_ARM64_FEATURE_ISA_PMULL == ZX_ARM64_FEATURE_ISA_PMULL {
-            features |= PMULL.mask;
-        }
-        if caps & ZX_ARM64_FEATURE_ISA_SHA2 == ZX_ARM64_FEATURE_ISA_SHA2 {
-            features |= SHA256.mask;
+        mod detect {
+            pub const FORCE_DYNAMIC_DETECTION: u32 = 0;
+            pub fn detect_features() -> u32 { 0 }
         }
     }
-
-    features
-}
-
-#[cfg(all(target_os = "windows", target_arch = "aarch64"))]
-fn detect_features() -> u32 {
-    // We do not need to check for the presence of NEON, as Armv8-A always has it
-    const _ASSERT_NEON_DETECTED: () = assert!((ARMCAP_STATIC & NEON.mask) == NEON.mask);
-    let mut features = 0;
-
-    let result = unsafe {
-        windows_sys::Win32::System::Threading::IsProcessorFeaturePresent(
-            windows_sys::Win32::System::Threading::PF_ARM_V8_CRYPTO_INSTRUCTIONS_AVAILABLE,
-        )
-    };
-
-    if result != 0 {
-        // These are all covered by one call in Windows
-        features |= AES.mask;
-        features |= PMULL.mask;
-        features |= SHA256.mask;
-    }
-
-    features
-}
-
-#[cfg(all(
-    target_arch = "aarch64",
-    any(target_os = "ios", target_os = "macos", target_os = "tvos")
-))]
-fn detect_features() -> u32 {
-    // TODO(MSRV 1.64): Use `name: &core::ffi::CStr`.
-    fn detect_feature(name: &[u8]) -> bool {
-        use crate::polyfill;
-        use core::mem;
-        use libc::{c_char, c_int, c_void};
-
-        let nul_terminated = name
-            .iter()
-            .position(|&b| b == 0)
-            .map(|index| (index + 1) == name.len())
-            .unwrap_or(false);
-        if !nul_terminated {
-            return false;
-        }
-        let name = polyfill::ptr::from_ref(name).cast::<c_char>();
-
-        let mut value: c_int = 0;
-        let mut len = mem::size_of_val(&value);
-        let value_ptr = polyfill::ptr::from_mut(&mut value).cast::<c_void>();
-        // SAFETY: `name` is nul-terminated and it doesn't contain interior nul bytes. `value_ptr`
-        // is a valid pointer to `value` and `len` is the size of `value`.
-        let rc = unsafe { libc::sysctlbyname(name, value_ptr, &mut len, core::ptr::null_mut(), 0) };
-        // All the conditions are separated so we can observe them in code coverage.
-        if rc != 0 {
-            return false;
-        }
-        debug_assert_eq!(len, mem::size_of_val(&value));
-        if len != mem::size_of_val(&value) {
-            return false;
-        }
-        value != 0
-    }
-
-    // We do not need to check for the presence of NEON, as Armv8-A always has it
-    const _ASSERT_NEON_DETECTED: () = assert!((ARMCAP_STATIC & NEON.mask) == NEON.mask);
-
-    let mut features = 0;
-
-    // TODO(MSRV 1.64): Use `: &CStr = CStr::from_bytes_with_nul_unchecked`.
-    // TODO(MSRV 1.77): Use c"..." literal.
-    const SHA512_NAME: &[u8] = b"hw.optional.armv8_2_sha512\0";
-    if detect_feature(SHA512_NAME) {
-        features |= SHA512.mask;
-    }
-
-    features
-}
-
-#[cfg(all(
-    any(target_arch = "aarch64", target_arch = "arm"),
-    not(any(
-        target_os = "android",
-        target_os = "fuchsia",
-        all(target_os = "linux", not(target_env = "uclibc")),
-        target_os = "windows",
-        any(target_os = "ios", target_os = "macos", target_os = "tvos"),
-    ))
-))]
-fn detect_features() -> u32 {
-    0
 }
 
 macro_rules! features {
@@ -252,7 +79,7 @@ macro_rules! features {
         )+
 
         // See const assertions below.
-        const ARMCAP_STATIC: u32 = ARMCAP_STATIC_DETECTED & ARMCAP_STATIC_FILTER;
+        const ARMCAP_STATIC: u32 = ARMCAP_STATIC_DETECTED & !detect::FORCE_DYNAMIC_DETECTION;
         const ARMCAP_STATIC_DETECTED: u32 = 0
             $(
                 | (
@@ -264,11 +91,6 @@ macro_rules! features {
                     }
                 )
             )+;
-        const ARMCAP_STATIC_FILTER: u32 = if cfg!(all(target_arch = "aarch64", target_os = "macos")) {
-            _AARCH64_APPLE_FEATURES
-        } else {
-            !0
-        };
 
         const ALL_FEATURES: [Feature; 5] = [
             $(
@@ -334,7 +156,7 @@ features! {
 //   `INIT`.
 // - See the safety invariants of `OPENSSL_armcap_P` below.
 pub unsafe fn init_global_shared_with_assembly() {
-    let detected = detect_features();
+    let detected = detect::detect_features();
     let filtered = (if cfg!(feature = "unstable-testing-arm-no-hw") {
         ALL_FEATURES
             .iter()
@@ -377,41 +199,13 @@ prefixed_extern! {
     static mut OPENSSL_armcap_P: u32;
 }
 
-// MSRV: Enforce 1.61.0 on some aarch64-* targets (aarch64-apple-*, in particular) prior to. Earlier
-// versions of Rust before did not report the AAarch64 CPU features correctly for these targets.
-// Cargo.toml specifies `rust-version` but versions before Rust 1.56 don't know about it.
-//
-// ```
-// $ rustc +1.61.0 --print cfg --target=aarch64-apple-ios | grep -E "neon|aes|sha|pmull"
-// target_feature="aes"
-// target_feature="neon"
-// target_feature="sha2"
-// $ rustc +1.61.0 --print cfg --target=aarch64-apple-darwin | grep -E "neon|aes|sha|pmull"
-// target_feature="aes"
-// target_feature="neon"
-// target_feature="sha2"
-// target_feature="sha3"
-//
-// XXX/TODO(coverage)/TODO(size): aarch64-apple-darwin is statically guaranteed to have "sha3" but
-// other aarch64-apple-* targets require dynamic detection. Since we don't have test coverage for
-// the other targets yet, we wouldn't have a way of testing the dynamic detection if we statically
-// enabled `SHA512` for -darwin. So instead, temporarily, we statically ignore the static
-// availability of the feature on -darwin so that it runs the dynamic detection.
-// ```
 #[allow(clippy::assertions_on_constants)]
 const _AARCH64_HAS_NEON: () =
     assert!(((ARMCAP_STATIC & NEON.mask) == NEON.mask) || !cfg!(target_arch = "aarch64"));
-const _AARCH64_APPLE_FEATURES: u32 = NEON.mask | AES.mask | SHA256.mask | PMULL.mask;
+
 #[allow(clippy::assertions_on_constants)]
-const _AARCH64_APPLE_TARGETS_EXPECTED_FEATURES: () = assert!(
-    ((ARMCAP_STATIC & _AARCH64_APPLE_FEATURES) == _AARCH64_APPLE_FEATURES)
-        || !cfg!(all(target_arch = "aarch64", target_vendor = "apple"))
-);
-#[allow(clippy::assertions_on_constants)]
-const _AARCH64_APPLE_DARWIN_TARGETS_EXPECTED_FEATURES: () = assert!(
-    (ARMCAP_STATIC == _AARCH64_APPLE_FEATURES)
-        || !cfg!(all(target_arch = "aarch64", target_os = "macos"))
-);
+const _FORCE_DYNAMIC_DETECTION_HONORED: () =
+    assert!((ARMCAP_STATIC & detect::FORCE_DYNAMIC_DETECTION) == 0);
 
 #[cfg(test)]
 mod tests {
