+++ src/cpu/arm/fuchsia.rs
@@ -1,0 +1,55 @@
+// Copyright 2016-2024 Brian Smith.
+//
+// Permission to use, copy, modify, and/or distribute this software for any
+// purpose with or without fee is hereby granted, provided that the above
+// copyright notice and this permission notice appear in all copies.
+//
+// THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHORS DISCLAIM ALL WARRANTIES
+// WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+// MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY
+// SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+// WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION
+// OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
+// CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+
+use super::{AES, ARMCAP_STATIC, NEON, PMULL, SHA256};
+
+pub const FORCE_DYNAMIC_DETECTION: u32 = 0;
+
+pub fn detect_features() -> u32 {
+    type zx_status_t = i32;
+
+    #[link(name = "zircon")]
+    extern "C" {
+        fn zx_system_get_features(kind: u32, features: *mut u32) -> zx_status_t;
+    }
+
+    const ZX_OK: i32 = 0;
+    const ZX_FEATURE_KIND_CPU: u32 = 0;
+    const ZX_ARM64_FEATURE_ISA_ASIMD: u32 = 1 << 2;
+    const ZX_ARM64_FEATURE_ISA_AES: u32 = 1 << 3;
+    const ZX_ARM64_FEATURE_ISA_PMULL: u32 = 1 << 4;
+    const ZX_ARM64_FEATURE_ISA_SHA2: u32 = 1 << 6;
+
+    let mut caps = 0;
+    let rc = unsafe { zx_system_get_features(ZX_FEATURE_KIND_CPU, &mut caps) };
+
+    let mut features = 0;
+
+    // We do not need to check for the presence of NEON, as Armv8-A always has it
+    const _ASSERT_NEON_DETECTED: () = assert!((ARMCAP_STATIC & NEON.mask) == NEON.mask);
+
+    if rc == ZX_OK {
+        if caps & ZX_ARM64_FEATURE_ISA_AES == ZX_ARM64_FEATURE_ISA_AES {
+            features |= AES.mask;
+        }
+        if caps & ZX_ARM64_FEATURE_ISA_PMULL == ZX_ARM64_FEATURE_ISA_PMULL {
+            features |= PMULL.mask;
+        }
+        if caps & ZX_ARM64_FEATURE_ISA_SHA2 == ZX_ARM64_FEATURE_ISA_SHA2 {
+            features |= SHA256.mask;
+        }
+    }
+
+    features
+}
